# Improve List performance in IEnumerable.SequenceEqual