@EgorBo 2022-04-07
Nice! I assume this change slightly regresses e.g. arrays of floats