@Wraith2 2022-04-08
Is removal of a generic constraint on a public method a breaking change? It can cause a callsite to become ambiguous where it was not before if there was a user provided extension that did not have the constraint.