@stephentoub 2022-04-08
> Is removal of a generic constraint on a public method a breaking change?

The method the constraint is being removed from isn't public.

> I assume this change slightly regresses

I'm more concerned this is going to regress the performance of arrays of reference types.  Prior to this change, they could take the fast path that uses MemoryExtensions.SequenceEqual on spans, avoiding all enumeration and interface dispatch costs.  With this change, they will be forced down the `IList<T>` code path.

@yesmey, can you please share benchmark numbers for reference types?