@yesmey 2022-04-09
@stephentoub you're correct, I missed checking reference types for arrays its the regression is pretty awful

Here's the same benchmark program as before but with string:
|         Method |        Job |                                                                                                          Toolchain | Length | IsArray |        Mean |     Error |    StdDev |      Median | Ratio | RatioSD |
|--------------- |----------- |------------------------------------------------------------------------------------------------------------------- |------- |-------- |------------:|----------:|----------:|------------:|------:|--------:|
| **SequenceEquals** | **Job-OPOOIY** | **main** |      **1** |    **True** |    **32.39 ns** |  **0.542 ns** |  **0.507 ns** |    **32.64 ns** |  **1.00** |    **0.00** |
| SequenceEquals | Job-JHKYCQ |   sequenceequals |      1 |    True |    43.76 ns |  0.017 ns |  0.013 ns |    43.76 ns |  1.35 |    0.02 |
|                |            |                                                                                                                    |        |         |             |           |           |             |       |         |
| **SequenceEquals** | **Job-OPOOIY** | **main** |      **4** |    **True** |    **52.10 ns** |  **0.800 ns** |  **0.709 ns** |    **52.28 ns** |  **1.00** |    **0.00** |
| SequenceEquals | Job-JHKYCQ |   sequenceequals |      4 |    True |    82.14 ns |  0.770 ns |  0.720 ns |    81.88 ns |  1.58 |    0.02 |
|                |            |                                                                                                                    |        |         |             |           |           |             |       |         |
| **SequenceEquals** | **Job-OPOOIY** | **main** |     **32** |    **True** |   **229.02 ns** |  **4.118 ns** |  **3.852 ns** |   **229.58 ns** |  **1.00** |    **0.00** |
| SequenceEquals | Job-JHKYCQ |   sequenceequals |     32 |    True |   445.67 ns |  0.498 ns |  0.466 ns |   445.65 ns |  1.95 |    0.03 |
|                |            |                                                                                                                    |        |         |             |           |           |             |       |         |
| **SequenceEquals** | **Job-OPOOIY** | **main** |    **256** |    **True** | **2,477.34 ns** | **46.960 ns** | **41.629 ns** | **2,496.02 ns** |  **1.00** |    **0.00** |
| SequenceEquals | Job-JHKYCQ |   sequenceequals |    256 |    True | 3,888.12 ns |  0.773 ns |  0.646 ns | 3,888.04 ns |  1.57 |    0.03 |

I'll see if I can make it work, but meanwhile we can just put the PR on hold